ENG-3347 Adds computing thresholds off of alloy xval
Using the confidence thresholds from alloy cross validation, I compute
the average of the best value chosen for predicting over a fold.

I believe the average of each fold makes the most sense, as compared
to computing a global optimum. This adds a new metric type, methods
and tests for this new functionality.