Implement content type detection
Fill in the skeleton of the ContentTypeDetector
FeatureTransformer to detect between HTML, XML,
or plain text (default) documents.

The detection is very basic at this point, first
checking the content string for an xml or html starting
tag. Second is for the existence of certain metadata
fields (from datasift) that signify document types we
know to only ever return html. If nothing is detected,
default is to return a plaintext contenttype.