Add PredictOptions to return tokens & tags
This adds the option to PredictOptions to return tokens and/or tags. This
will allow use to know what tokens were in a span, and what tags were
assigned to those tokens. This change is motivated primarily for the need
to be able to do token level evaluation, rather than span level.