Fixes recall calculation at the span metrics level
I was losing counts of gold items because I was doing a flatmap on
a map, which turned it into a map, and not a sequence like I wanted.
Ironically I put a comment above warning me of this, but somehow at
some point I changed the code there...

So now added unit test to make sure this doesn't happen.