ENG-3367 Adds Ensemble Model
Adds Ensemble Model & SpanEnsembleModel.
The idea is that ensemble model is abstract and will have
two implementations: one for spans the other for classification.
Most of the code and logic should be in EnsembleModel if
it is shared, otherwise it delegates to the subclasses.

This includes an incomplete SpanEnsembleModel stub.