Add unicode conversions for spans
With the going back and forths between jruby and scala
code, span annotation and prediction representations
need to be quickly translated from utf-32/code points
as used in ididat/idibin, to utf-16 as used in idiml
(and vice versa).

The UnicodeConverter class contains a toIdidat and toIdiml
method for respectively converting the span offset and length
values into the unicode version of that project.