ENG-3367 Adds Static Methods for Rule Span Merging Logic
This adds methods to deal with cases when spans overlap.
They can overlap if different rules for different labels produce
overlapping spans, as well as predicted spans overlapping with
rules ones.

Primarily we deal with it by:
 1) if any rules overlaps with a predicted span, take the rule span
 2) if any rules spans overlap, take the one with the largest weight
      away from 0.5, else do some tie breaking on length and label name.

We could have gone a route to try to construct some "optimal set" of
spans from a set of overlapping spans, but felt that greedy reduction
should suffice.

Includes unit tests.