Adds type tag to BasicAlloyForge and wraps models in ensemble model
To be able to accommodate rules & predictions, we need the basic alloy
forge to wrap the models returned in an enesemble model. However the
ensemble model class is typed, and thus we use type tags to know which
kind of ensemble model to create.

This then means the alloy forge spec needed to be updated to reflect
this; I removed the JunkResult and based the test on Spans.