Adds labels and rules to train options
This adds a function to add rules to training options, as well
as exposing labels in training options. Fixes up cross validating
alloy forge to create new training options correctly.

This will be used by the SpanRuleFurnace to map
UUIDs to Human Labels to make named groups easier
to work with.