Fixes up stopword transformer loading
This adds the ability to load a transformer from the resources
for training, and then loading them from an alloy for prediction.
Adds a unit test for saving & loading, also creates analogous test
resources so that the unit tests work too.

The idea here is that we only save custom stopwords to the alloy
that differ from our default set of stopwords which will live in
the jar that the stopwords transformer depends on.

Also we only for now support loading custom stopwords from a local
file, as opposed to a web or S3 resoruce, etc.

Note: stopword files require the precomposed versions of letters,
that way when we normalize them, we'll also include the decomposed
variant.