ENG-936 Adds utility to output top n-grams
This checks in a new app that instantiates a feature pipeline
and outputs a csv of n-gram counts. It defaults to using a
configuration that lower cases everything and keeps words & punctuation.

The idea is that we basically featurize the data, sum over the dimensions,
sort by the counts and then produce the appropriate feature.

You can pass in your own feature pipeline configuration if you wanted
it to pass through a stopword filter, remove punctuation etc.