Adds competitive alloy forge
This adds the analogous forge that exists in the old system but
using the new way of doing things.

This will allow you to make different alloy configurations
compete with each other, and then having the winner chosen.
You do this all through the configuration.

Adds unit test on finding the max by F1.