ENG-1593 Adds TF-IDF Transformer
This can be used instead of the IndexTransformer. Rather than outputting
term frequencies it outputs tf-idf values for that term.

This adds a trait for calculating IDF values & follows the vocabulary
& index transformer pattern of having a mutable and immutable
object. You can have a frozen mutable object, likewise you can
also only create a immutable object by saving a mutable one.

This also fixes a config bug on the index transformer, since it
is now robust to configuration not being there.