Adds hook to add default stopword language
If a language is not detected in the content, the current fallback was to an empty string,
and thus an empty stopword set. With this change, it allows you, the model building magician, to
set a custom stopword list for things that the language detector is confused about for
instance. It also allows (the more likely usecase) falling back to a default language stopword set,
if that is what you so desire. Otherwise if you don't need this, the current behavior is left
unchanged.